[FEAT] - Managing quantity of items

+++ src/contexts/CartProvider.jsx
@@ -4,21 +4,32 @@ export const CartContext = createContext(null)
 
 export function CartProvider(props) {
     const [item, setItem] = useState([]);
+    const [inputQuantity, setInputQuantity] = useState(1);
 
-    const AddItemToCart = (product) => {
-        setItem(prev => [...prev, product]);
-        console.log("adicionado")
+    const addItemToCart = (product) => {
+        const itemExists = item.find(currentProduct => currentProduct.id === product.id);
+        if(!itemExists) {
+            setItem(prev => [...prev, product]);
+            console.log("Adicionado")
+        } else {
+            console.log("Produto já adicionado no carrinho")
+        }
     }
 
-    const RemoveItemFromList = (productId) => {
+    const removeItemFromList = (productId) => {
         const filteredRemovedItems = item.filter(currentItem => {
             return currentItem.id !== productId
         });
         setItem(filteredRemovedItems); 
     }
 
+    const updateItemQuantity = () => {
+        setInputQuantity(prevQuantity => prevQuantity + 1)
+        console.log("clicou")
+    }
+
     return  (
-        <CartContext.Provider value={{ item, AddItemToCart, RemoveItemFromList }}>
+        <CartContext.Provider value={{ item, addItemToCart, removeItemFromList, inputQuantity, updateItemQuantity }}>
             {props.children} 
         </CartContext.Provider>
     )
